+++ packages/static/src/staticFiles.ts
@@ -15,7 +15,7 @@ export default async function staticFiles(
   const {
     pages,
     router: { routes, options: routerOptions = () => ({}) },
-    output: { render, dir, redirects = false },
+    output: { render, dir },
     history: historyOptions
   } = config;
 
@@ -63,14 +63,6 @@ export default async function staticFiles(
             (emitted: Emitted) => {
               try {
                 const { response } = emitted;
-                if (response.redirect && !redirects) {
-                  resolve({
-                    pathname,
-                    success: false,
-                    error: new Error("redirect")
-                  });
-                  return;
-                }
                 const html = render(emitted);
                 fs.outputFile(outputPath, html).then(() => {
                   resolve({ pathname, success: true });