+++ packages/static/src/types.ts
@@ -17,7 +17,6 @@ export type GetRouterOptions = () => RouterOptions;
 export interface StaticOutput {
   render: (emitted: Emitted) => string;
   dir: string;
-  redirects?: boolean;
 }
 
 export interface StaticRouter {