+++ packages/static/CHANGELOG.md
@@ -1,6 +1,7 @@
 ## Next
 
 * Remove `insert` option to `staticFiles`. `render` should return an HTML string.
+* Remove `redirects` option. `render` should throw for routes that shouldn't be rendered.
 
 ## 2.0.0-alpha.0